Refactoring: states and transitions are now named using a UniqueIdProvider.
The current implementation of UniqueIdProvider is based on uuid4, which provides random strings on demand.

A bug remains in Before(True, True), which is not systematically caught by the provided unit tests, probably because the relative order of uuids vary from a run to an other. Must be fixed in the future.