Implement a mechanism for sending a MetaEvent from Statechart Code Blocks.
    meta("metadata", datum="foo", value="bar)

To support Distributed Statecharts, where one Statechart would like to
communicate Meta-Information to other Statecharts which are possibly
running as separate Systems (i.e running in other Processes, Containers
or Servers).

For Distributed Statecharts these MetaEvents are routed to "Adapters"
which implement the Property Statechart Interface. These Adapters are
attached to the Statechart Interpreter in the normal way. The
MetaEvents are intercepted by an Adapter and communicated to the other
Statecharts accordign to the methods defined in each particular Adapter.

The MetaEvents are not sent to the internal Event Queue and they should
not interfear with the existing operation of Property Statecharts as
these MetaEvents are identifed in Property Statecharts as
"meta event sent" rather than "event sent" (as seen when send() is used).