Refactor code/python.py into two classes:
  code/enhanced.py - implements enhanced evaluator functionality.
  code/python.py - implements the Python code evaluation.

It's now possible to create a Code Evaluator which inherits from
either the Evaluator or EnhancedEvaluator Class, depending on the
desired behaviour required for the Code Evaluator.

Eventually the EnhancedEvaluator Class may be converted to a Mixin
Class so that additional behaviours can be added to an Evaluator
without needing to refactor exising API interfaces.